fix cyclic deps and add a tslint rule to catch this on presubmit
